v0.97 - Release Candidate 3 - Настройки режима динамического изменения частоты для драйвера RMT (постоянная частота процессора и APB, если драйвер инициализирован). На текущий момент функционал отключен. Включается через menuconfig - Обновлен код в связи с обновлением ESP IDF и его компонентов - Изменены настройки буферов WiFi (бОльшая стабильность подключения к Wi-Fi сетям) - Скорректировано поведение прерывания индикатора из-за обновившихся правил FreeRTOS asserts - Изменено возвращаемое значение сенсора IR, добавлено отображение сырых таймингов при запросе

+++ src/include/Settings.h
@@ -35,7 +35,7 @@ class Settings_t {
 		} TimeSync;
 
 		struct WiFi_t {
-			const uint16_t					IPCountdown 	= 100000;
+			const uint32_t					IPCountdown 	= 100000;
 			const uint8_t					SavedAPCount	= 8;
 			const string 					DefaultIP		= "192.168.4.1";
